+++ tests/test_gitlab_create_mr.py
@@ -142,7 +142,7 @@ class TestGitLabCreateMR(BaseTest):
                 |
                 x-drop-constraint (untracked)
 
-            Fetching origin...
+            Checking if target branch ignore-trailing exists on origin remote... YES
             Creating a draft MR from chore/fields to ignore-trailing... OK, see www.gitlab.com
             Checking for open GitLab MRs (to determine MR chain)... OK
             Updating description of MR !5 to include the chain of MRs... OK
@@ -288,7 +288,7 @@ class TestGitLabCreateMR(BaseTest):
               |
               o-testing/endpoints
 
-            Fetching origin...
+            Checking if target branch develop exists on origin remote... YES
             Creating a MR from allow-ownership-link to develop... OK, see www.gitlab.com
             Setting milestone of MR !7 to 42... OK
             Adding gitlab_user as assignee to MR !7... OK
@@ -424,8 +424,7 @@ class TestGitLabCreateMR(BaseTest):
             """
         rewrite_branch_layout_file(body)
 
-        expected_msg = ("Fetching origin...\n"
-                        "Warn: target branch for this MR (feature/api_handling) is not found on remote, pushing...\n"
+        expected_msg = ("Checking if target branch feature/api_handling exists on origin remote... NO\n"
                         "Pushing untracked branch feature/api_handling to origin...\n"
                         "Creating a MR from feature/api_exception_handling to feature/api_handling... OK, see www.gitlab.com\n")
 
@@ -560,7 +559,7 @@ class TestGitLabCreateMR(BaseTest):
             |
             o-feature *
 
-        Fetching origin_1...
+        Checking if target branch branch-1 exists on origin_1 remote... YES
         Creating a MR from feature to branch-1... OK, see www.gitlab.com
         Checking for open GitLab MRs (to determine MR chain)... OK
         Updating description of MR !16 to include the chain of MRs... OK
@@ -611,7 +610,7 @@ class TestGitLabCreateMR(BaseTest):
         expected_result = """
         Add feature_1 onto the inferred upstream (parent) branch feature? (y, N)
         Added branch feature_1 onto feature
-        Fetching origin_2...
+        Checking if target branch feature exists on origin_2 remote... YES
         Creating a MR from feature_1 to feature... OK, see www.gitlab.com
         Checking for open GitLab MRs (to determine MR chain)... OK
         Updating description of MR !17 to include the chain of MRs... OK
@@ -649,7 +648,7 @@ class TestGitLabCreateMR(BaseTest):
               |
               o-feature_2 *
 
-        Fetching origin_1...
+        Checking if target branch feature exists on origin_1 remote... YES
         Creating a MR from feature_2 to feature... OK, see www.gitlab.com
         Checking for open GitLab MRs (to determine MR chain)... OK
         Updating description of MR !18 to include the chain of MRs... OK
@@ -671,7 +670,7 @@ class TestGitLabCreateMR(BaseTest):
         expected_result = """
         Add feature_3 onto the inferred upstream (parent) branch feature_2? (y, N)
         Added branch feature_3 onto feature_2
-        Fetching origin_1...
+        Checking if target branch feature_2 exists on origin_1 remote... YES
         Creating a MR from feature_3 to feature_2... OK, see www.gitlab.com
         Checking for open GitLab MRs (to determine MR chain)... OK
         Updating description of MR !19 to include the chain of MRs... OK
@@ -693,8 +692,7 @@ class TestGitLabCreateMR(BaseTest):
         expected_result = """
         Add feature_4 onto the inferred upstream (parent) branch feature_3? (y, N)
         Added branch feature_4 onto feature_3
-        Fetching origin_2...
-        Warn: target branch for this MR (feature_3) is not found on remote, pushing...
+        Checking if target branch feature_3 exists on origin_2 remote... NO
         Push untracked branch feature_3 to origin_2? (y, Q)
         Creating a MR from feature_4 to feature_3... OK, see www.gitlab.com
         """
@@ -716,8 +714,7 @@ class TestGitLabCreateMR(BaseTest):
         expected_result = """
         Add feature_5 onto the inferred upstream (parent) branch feature_3? (y, N)
         Added branch feature_5 onto feature_3
-        Fetching origin...
-        Warn: target branch for this MR (feature_3) is not found on remote, pushing...
+        Checking if target branch feature_3 exists on origin remote... NO
         Push untracked branch feature_3 to origin? (y, Q)
         Creating a MR from feature_5 to feature_3... OK, see www.gitlab.com
         """
@@ -742,7 +739,7 @@ class TestGitLabCreateMR(BaseTest):
         assert_success(
             ['gitlab', 'create-mr'],
             """
-            Fetching origin...
+            Checking if target branch master exists on origin remote... YES
             Creating a MR from develop to master... OK, see www.gitlab.com
             """
         )
@@ -787,7 +784,7 @@ class TestGitLabCreateMR(BaseTest):
             """
             Warn: Branch develop is behind its remote counterpart. Consider using git pull.
             Proceed with creating merge request? (y, Q)
-            Fetching origin...
+            Checking if target branch master exists on origin remote... YES
             Creating a MR from develop to master... OK, see www.gitlab.com
             """
         )
@@ -847,7 +844,56 @@ class TestGitLabCreateMR(BaseTest):
             """
             Warn: Branch develop is diverged from and older than its remote counterpart. Consider using git reset --keep.
             Proceed with creating merge request? (y, Q)
-            Fetching origin...
+            Checking if target branch master exists on origin remote... YES
             Creating a MR from develop to master... OK, see www.gitlab.com
+            """
+        )
+
+    def test_gitlab_create_mr_when_target_branch_disappeared_from_remote(self, mocker: MockerFixture) -> None:
+        self.patch_symbol(mocker, 'git_machete.code_hosting.OrganizationAndRepository.from_url', mock_from_url)
+        self.patch_symbol(mocker, 'git_machete.gitlab.GitLabToken.for_domain', mock_gitlab_token_for_domain_none)
+        self.patch_symbol(mocker, 'urllib.request.urlopen', mock_urlopen(MockGitLabAPIState()))
+
+        (
+            self.repo_sandbox
+            .new_branch("develop").commit().push()
+            .new_branch("feature").commit().push()
+        )
+        (
+            self.repo_sandbox
+            .chdir(self.repo_sandbox.remote_path)
+            .delete_branch("develop")
+            .chdir(self.repo_sandbox.local_path)
+        )
+
+        rewrite_branch_layout_file("develop\n\tfeature")
+        self.patch_symbol(mocker, 'builtins.input', mock_input_returning('y'))
+        assert_success(
+            ['gitlab', 'create-mr'],
+            """
+            Checking if target branch develop exists on origin remote... NO
+            Push untracked branch develop to origin? (y, Q)
+            Creating a MR from feature to develop... OK, see www.gitlab.com
+            """
+        )
+
+    def test_gitlab_create_mr_when_target_branch_appeared_on_remote(self, mocker: MockerFixture) -> None:
+        self.patch_symbol(mocker, 'git_machete.code_hosting.OrganizationAndRepository.from_url', mock_from_url)
+        self.patch_symbol(mocker, 'git_machete.gitlab.GitLabToken.for_domain', mock_gitlab_token_for_domain_none)
+        self.patch_symbol(mocker, 'urllib.request.urlopen', mock_urlopen(MockGitLabAPIState()))
+
+        (
+            self.repo_sandbox
+            .new_branch("develop").commit().push()
+            .delete_remote_branch("origin/develop")
+            .new_branch("feature").commit().push()
+        )
+
+        rewrite_branch_layout_file("develop\n\tfeature")
+        assert_success(
+            ['gitlab', 'create-mr'],
+            """
+            Checking if target branch develop exists on origin remote... YES
+            Creating a MR from feature to develop... OK, see www.gitlab.com
             """
         )